requirements.txt
httpx
rq
starlette
uvicorn
jinja2
beautifulsoup4
motor
aiofiles
python-slugify
